refactor: change file name(random file) and class constructor
